add  update file by 20200113

+++ intelligencecity_manager_interface/src/main/java/com/shanghaiuniversity/manager/service/AdministratoruserService.java
@@ -42,14 +42,14 @@ public interface AdministratoruserService {
      * @param id
      * @return
      */
-    public Administratoruser findOne(Long id);
+    public Administratoruser findOne(int id);
 
     /**
      * 批量删除
      *
      * @param ids
      */
-    public void delete(Long[] ids);
+    public void delete(int[] ids);
 
     /**
      * 分页
